requirements.txt
appdirs==1.4.3
asgiref==3.2.5
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.40.0
django-appconf==1.0.3
django-ckeditor==5.7.1
django-cors-headers==3.1.1
django-filter==2.2.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==3.2.0
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
drf-writable-nested==0.5.4
idna==2.8
jsonfield==2.0.2
Markdown==3.1.1
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.7.0
pilkit==2.0
Pillow==6.2.1
python3-openid==3.1.0
pytz==2019.3
regex==2020.2.20
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.6
wincertstore==0.2